+++ lib/constants/app_info.dart
@@ -1,8 +1,8 @@
 /// App version and info constants
 /// Update version here only - all other files will reference this
 class AppInfo {
-  static const String version = '3.5.1';
-  static const String buildNumber = '75';
+  static const String version = '3.5.2';
+  static const String buildNumber = '76';
   static const String fullVersion = '$version+$buildNumber';