+++ lib/l10n/app_localizations_de.dart
@@ -2450,6 +2450,15 @@ class AppLocalizationsDe extends AppLocalizations {
   String get libraryFilterDateYear => 'This Year';
 
   @override
+  String get libraryFilterSort => 'Sort';
+
+  @override
+  String get libraryFilterSortLatest => 'Latest';
+
+  @override
+  String get libraryFilterSortOldest => 'Oldest';
+
+  @override
   String libraryFilterActive(int count) {
     return '$count filter(s) active';
   }