+++ lib/l10n/app_localizations_id.dart
@@ -2448,6 +2448,15 @@ class AppLocalizationsId extends AppLocalizations {
   String get libraryFilterDateYear => 'This Year';
 
   @override
+  String get libraryFilterSort => 'Sort';
+
+  @override
+  String get libraryFilterSortLatest => 'Latest';
+
+  @override
+  String get libraryFilterSortOldest => 'Oldest';
+
+  @override
   String libraryFilterActive(int count) {
     return '$count filter(s) active';
   }