+++ lib/l10n/app_localizations_ru.dart
@@ -2481,6 +2481,15 @@ class AppLocalizationsRu extends AppLocalizations {
   String get libraryFilterDateYear => 'This Year';
 
   @override
+  String get libraryFilterSort => 'Sort';
+
+  @override
+  String get libraryFilterSortLatest => 'Latest';
+
+  @override
+  String get libraryFilterSortOldest => 'Oldest';
+
+  @override
   String libraryFilterActive(int count) {
     return '$count filter(s) active';
   }